refactor(mailer): consistent naming
