+++ internals/isc/src/mailer/index.js
@@ -1,3 +1,3 @@
-export * from './subscription/index.js';
+export * from './subscriptions/index.js';
 export * from './errors.js';
-export * from './mail/index.js';
+export * from './mails/index.js';
