+++ internals/isc/src/mailer/subscriptions/http.js
@@ -3,19 +3,19 @@ import { createSignedGetRequestFactory, createSignedRequestFactory } from '../..
 /**
  * @type {import('../../common/types.js').CommonGetRequestFactory<true>}
  */
-export const createGetSubscriptionRequest = createSignedGetRequestFactory('/subscription', {
+export const createGetSubscriptionRequest = createSignedGetRequestFactory('/subscriptions', {
   token: true,
 });
 
 /**
  * @type {import('../../common/types.js').CommonRequestFactory<import('./dto.js').CreateSubscriptionRequestDTO>}
  */
-export const createSubscriptionRequest = createSignedRequestFactory('/subscription');
+export const createSubscriptionRequest = createSignedRequestFactory('/subscriptions');
 
 /**
  * @type {import('../../common/types.js').CommonRequestFactory<import('./dto.js').UpdateDomainSubscriptionRequestDTO, true>}
  */
-export const createUpdateDomainSubscriptionRequest = createSignedRequestFactory('/subscription', {
+export const createUpdateDomainSubscriptionRequest = createSignedRequestFactory('/subscriptions', {
   method: 'PATCH',
   token: true,
 });
