+++ services/mailer/src/routes/mails/[token]/+server.ts
